+++ forui/CHANGELOG.md
@@ -14,6 +14,8 @@
 * Add `FTimePickerStyle`.
 * Add `FPickerStyle.selectionHeightAdjustment`.
 * Add `FTileGroup.physics`.
+* Add `FSelectTileGroup.physics`.
+* Add `FSelectMenuTile.physics`.
 
 ### Changes